remove old ContactForm and bring in new component
